Minor change in the notebook tests.
Versions of Python greater than 3.7 (or 3.8?) may throw a TypeError when
running test_all_notebooks.py. It's out of the scope of what I'm doing
here, but I made a note in the code in case the issue comes up as
landlab and terrainbento are upgraded to future versions of python.