Set weathering rate to a very small but nonzero value
In the tests for the HySa and SaVs models, the soil weathering rate
is set to zero. This value is passed to the Landlab ExponentialWeatherer,
which raises an error (on line 153) because a nonzero value is expected.
I changed the test weathering rate to be a very small but nonzero value
for it to work with the current version of ExponentialWeatherer.