Support developer install with current Python and Landlab versions (#159)
* Fix yaml formatting issue

* Replace versioneer with zest.releaser

* Match keywords with the current version of ExponentialWeatherer

The current version of ExponentialWeatherer uses only one underscore
in its keyword arguments instead of two. I updated the models that
used the old syntax to use the new.

* Set weathering rate to a very small but nonzero value

In the tests for the HySa and SaVs models, the soil weathering rate
is set to zero. This value is passed to the Landlab ExponentialWeatherer,
which raises an error (on line 153) because a nonzero value is expected.
I changed the test weathering rate to be a very small but nonzero value
for it to work with the current version of ExponentialWeatherer.

* Add additional digits of precision to truth file

This allows tests to pass with the current versions of Python (3.12)
and Landlab (2.8).

* Remove encoding keyword to nbformat.read

* Fix typo